First implementation of object tracking.
This is done by extending a protocol message to include an array of object IDs passed outside of the payload, so that Twibd can pick them up.
Twibd will keep track of which objects a client owns, and when the client is destroyed, those objects will be destroyed if the client didn't
close them itself. One change that this made necessary was that Request objects in Twibd now carry a strong pointer to the client that created
them, to keep the client alive until all of its requests have been dispatched. This is important because a client may dispatch several close
requests, then mark itself for deletion. The socket thread would previously have destroyed the only reference, immediately causing Twibd to
queue close requests for the objects it thinks have been lost, but have just not been marked as cleaned up yet since the close requests have
not yet been dispatched.

On the Twili side of things, ResponseWriter has been given a method to add an object to the current message's list of object IDs. The number
of object IDs being sent, much like the payload size, must be known when the response is opened. A `Finalize` method has also been added, to
actually send the list of object IDs. This method will also ensure that the correct amount of data and correct number of object IDs have been
written, throwing an exception if either condition is not met. I'd do this in ResponseState::~ResponseState, but throwing exceptions in a
destructor is really hairy.

One more thing to note is that the messages that previously returned object IDs in their payloads now use indexes into the message's object list
in their place. This makes it very easy for Twili, Twibd, and Twib to all use RAII to keep track of object lifetimes.