twib: refactor NamedPipeServer to platform::windows::EventLoop
The new EventLoop is more general and is not tied to the named pipe use case. To facilitate this, "pipes" as they pertained to NamedPipeServer have been renamed to "members" (as in members of the event loop), and each NamedPipeMessageConnection now uses two members, one for input, and one for output. Members also now do not need to own their events, since we don't own OvlK events.