windows: make InputPump use synchronous I/O
Unfortunately, it seems like there's no winning with async I/O for console input on windows :'(
Stock console is a magic kind of handle and only kindof behaves like a File for the high-level
interface, but that doesn't support overlapped I/O. I implemented a sucky line driver and used
the low-level interface, which supported async but gave raw keyboard events, but that was sub-optimal
and didn't work with msys.

Solution is to do what .NET apparently does and use a background thread for read operations.
Fortunately, windows has CancelSynchronousIo which lets us wake up the background thread when we
need to kill it...