gdb: improve qOffsets for non-monitored processes
Now we count how many modules there are. If there's only one, we assume it's
main and pick it.  If there are more than one, we assume the first is rtld and
the second is main and pick main.

Also, we do the address space search on the device now. Games tend to have very
fragmented address spaces and it takes too long to round-trip for each
QueryMemory call.