bridge: Don't error out when sending a 0-byte response.
Both the USB and TCP bridge had issues in this case.  For USB, it would
try to send a 0-byte USB packet but the USB stack would return an error.
For TCP, it seems like bsd_send would return 0 and twili would then
interpret that as an error (haven't actually tested this).